benchmark: isolate FP8 MM performance test in dedicated module
